renamed + start on xml parsing
