+++ bpmnconstraints/parser/json_model.py
@@ -1,0 +1,6 @@
+class JsonProperties:
+    def __init__(self, model) -> None:
+        self.model = model
+
+    def get_child_shapes(self):
+        pass
