+++ bpmnconstraints/parser/xml_model.py
@@ -1,0 +1,9 @@
+class XmlProperties:
+    def __init__(self, model) -> None:
+        self.model = model
+
+    def get_child_shapes(self):
+        pass
+
+    def child_shapes():
+        pass
